empManagerView

+++ src/main/java/com/example/checkmateqq/EmpManagerViewController.java
@@ -1,18 +1,20 @@
 package com.example.checkmateqq;
 
+import com.example.checkmateqq.triedy.Shift;
+import com.example.checkmateqq.triedy.Uhs;
 import com.example.checkmateqq.triedy.User;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
-import javafx.scene.control.ListView;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
+import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.image.ImageView;
 import javafx.scene.input.InputMethodEvent;
 import javafx.scene.input.MouseEvent;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 public class EmpManagerViewController {
 
@@ -23,12 +25,17 @@ public class EmpManagerViewController {
     private ImageView IconSearch;
 
     @FXML
-    private ListView<?> pastShiftsList;
+    private ListView<String> pastShiftsList;
 
     @FXML
-    private ListView<?> upcomingShiftsList;
+    private ListView<String> upcomingShiftsList;
     private UserDao userDao = DaoFactory.INSTANCE.getUserDao();
+    private ShiftDao shiftDao = DaoFactory.INSTANCE.getShiftDao();
+    private StationDao stationDao = DaoFactory.INSTANCE.getStationDao();
+    private UhsDao uhsDao = DaoFactory.INSTANCE.getUhsDao();
+    private Map<Integer, String> upcomingShiftsMap = new HashMap<>();
     private int selectedId;
+
     @FXML
     private void initialize() throws EntityNotFoundException {
         employeeTableColumns();
@@ -40,9 +47,13 @@ public class EmpManagerViewController {
                         // Handle the selection change
                         selectedId = newValue.getId();
                         System.out.println("Selected ID: " + selectedId);
+                        fillPastShiftsList();
+                        fillUpcomingShiftsList();
                     }
                 });
+
     }
+
     @FXML
     void change(InputMethodEvent event) {
 
@@ -50,14 +61,50 @@ public class EmpManagerViewController {
 
     @FXML
     void onCancelShift(ActionEvent event) {
+        String selectedString = upcomingShiftsList.getSelectionModel().getSelectedItem();
+        int selectedShiftId = 0;
+        for (Map.Entry<Integer, String> entry : upcomingShiftsMap.entrySet()) {
+            Integer shiftId = entry.getKey();
+            String shiftString = entry.getValue();
+            if (selectedString.equals(shiftString)) {
+                selectedShiftId = shiftId;
+            }
+        }
+        uhsDao.deleteUhsByShiftIdAndUserId(selectedShiftId,selectedId);
+        upcomingShiftsList.getItems().clear();
+        fillUpcomingShiftsList();
 
     }
 
     @FXML
     void onDeleteAccount(ActionEvent event) throws EntityNotFoundException {
-        userDao.deleteUserById(selectedId);
+        if (userDao.hasUpcomingShifts(selectedId)) {
+            Alert alert = new Alert(Alert.AlertType.WARNING);
+            alert.setHeaderText(null);
+            alert.setContentText("Employee still has upcoming shifts.");
+            alert.show();
+        } else {
+            Alert confirmationAlert = new Alert(Alert.AlertType.CONFIRMATION);
+            confirmationAlert.setHeaderText("Are you sure you want to delete the user?");
+            confirmationAlert.setContentText("Click OK to proceed, or Cancel to abort.");
+            confirmationAlert.showAndWait().ifPresent(response -> {
+                if (response == ButtonType.OK) {
+                    System.out.println("User clicked OK. Proceeding...");
+                    try {
+                        userDao.deleteUserById(selectedId);
+                    } catch (EntityNotFoundException e) {
+                        throw new RuntimeException(e);
+                    }
+                } else if (response == ButtonType.CANCEL) {
+                    System.out.println("User clicked Cancel. Aborting...");
+                    confirmationAlert.close();
+                }
+            });
+        }
+
         employeeTable.getItems().clear();
         fillEmployeeTable();
+
     }
 
     @FXML
@@ -72,26 +119,61 @@ public class EmpManagerViewController {
 
     public void setUserId(User user) {
     }
-    private void fillEmployeeTable(){
-        for(User user: userDao.returnEmployees()){
+
+    private void fillEmployeeTable() {
+        for (User user : userDao.returnEmployees()) {
             Integer id = user.getId();
             String fullName = user.getName() + " " + user.getSurname();
-            EmployeeForTable employee = new EmployeeForTable(id,fullName);
+            EmployeeForTable employee = new EmployeeForTable(id, fullName);
             employeeTable.getItems().add(employee);
         }
 
     }
-    private void employeeTableColumns(){
+
+    private void employeeTableColumns() {
         employeeTable.setPrefWidth(190);
         TableColumn<EmployeeForTable, Integer> idColumn = new TableColumn<EmployeeForTable, Integer>("id");
         idColumn.setCellValueFactory(new PropertyValueFactory<EmployeeForTable, Integer>("id"));
         employeeTable.getColumns().add(idColumn);
         idColumn.setPrefWidth(80);
 
-        TableColumn<EmployeeForTable, String> nameColumn = new TableColumn<EmployeeForTable, String>("id");
+        TableColumn<EmployeeForTable, String> nameColumn = new TableColumn<EmployeeForTable, String>("Employee");
         nameColumn.setCellValueFactory(new PropertyValueFactory<EmployeeForTable, String>("fullName"));
         employeeTable.getColumns().add(nameColumn);
         nameColumn.setPrefWidth(157);
+    }
+
+    private void fillPastShiftsList() {
+        List<String> pastShiftsData = new ArrayList<>();
+        for (Shift shift : shiftDao.getFutureShiftsForUser(selectedId)) {
+            String cas = "";
+            if (shift.isFirst()) {
+                cas = "first";
+            } else if (!shift.isFirst()) {
+                cas = "second";
+            }
+            String shiftString = shift.getDate() + ", " + cas + ", " + (stationDao.getStationById(shift.getStation_id()).toString2());
+            pastShiftsData.add(shiftString);
+        }
+        pastShiftsList.getItems().setAll(pastShiftsData);
+    }
+
+
+    private void fillUpcomingShiftsList() {
+        upcomingShiftsMap.clear();
+        List<String> upcomingShiftsData = new ArrayList<>();
+        for (Shift shift : shiftDao.getFutureShiftsForUser(selectedId)) {
+            String cas = "";
+            if (shift.isFirst()) {
+                cas = "7:00  - 13:00";
+            } else if (!shift.isFirst()) {
+                cas = "13:00 - 19:00";
+            }
+            String shiftString = shift.getDate() + ", " + cas + ", " + (stationDao.getStationById(shift.getStation_id()).toString());
+            upcomingShiftsMap.put(shift.getId(), shiftString);
+            upcomingShiftsData.add(shiftString);
+        }
+        upcomingShiftsList.getItems().setAll(upcomingShiftsData);
     }
 }
